+++ statClient/src/components/table.tsx
@@ -30,43 +30,43 @@ interface IData {
 	epocas: number;
 	rank: number;
 }
-// Método en reemplazo del hook useSort, por medio del cual se ejecutaban un
-// ordenamiento según instrucciones, sobre el erray en bruto extraído
-// de la BD.
-function sortArray(dataToSort: any) {
-	if (dataToSort.length > 0) {
-		// Organización por menor tiempo que se hace primero para hacer el descarte
-		dataToSort.sort(function (a: any, b: any) {
-			return a.total_response_time - b.total_response_time;
-		});
-		//Organización por porcentaje de errores, que son los errores de un usuario
-    //con respecto a las respuestas respondidas
-		dataToSort.sort(function (a: any, b: any) {
-			return a.error_percentage - b.error_percentage;
-		});
-		//Organización por terminación del album (album finalizado o no)
-		dataToSort.sort(function (a: any, b: any) {
-			if (a.ended_album > b.ended_album) {
-				return -1;
-			}
+// // Método en reemplazo del hook useSort, por medio del cual se ejecutaban un
+// // ordenamiento según instrucciones, sobre el erray en bruto extraído
+// // de la BD.
+// function sortArray(dataToSort: any) {
+// 	if (dataToSort.length > 0) {
+// 		// Organización por menor tiempo que se hace primero para hacer el descarte
+// 		dataToSort.sort(function (a: any, b: any) {
+// 			return a.total_response_time - b.total_response_time;
+// 		});
+// 		//Organización por porcentaje de errores, que son los errores de un usuario
+// 		//con respecto a las respuestas respondidas
+// 		dataToSort.sort(function (a: any, b: any) {
+// 			return a.error_percentage - b.error_percentage;
+// 		});
+// 		//Organización por terminación del album (album finalizado o no)
+// 		dataToSort.sort(function (a: any, b: any) {
+// 			if (a.ended_album > b.ended_album) {
+// 				return -1;
+// 			}
 
-			if (a.ended_album < b.ended_album) {
-				return 1;
-			}
-		});
-	}
-	return addRank(dataToSort);
-}
-//Método que añade posiciones
-function addRank(arrayData: any) {
-	let position = 1;
-	for (let index = 0; index < arrayData.length; index++) {
-		const element = arrayData[index];
-		element.rank = position++;
-	}
-	// console.log(arrayData)
-	return arrayData;
-}
+// 			if (a.ended_album < b.ended_album) {
+// 				return 1;
+// 			}
+// 		});
+// 	}
+// 	return addRank(dataToSort);
+// }
+// //Método que añade posiciones
+// function addRank(arrayData: any) {
+// 	let position = 1;
+// 	for (let index = 0; index < arrayData.length; index++) {
+// 		const element = arrayData[index];
+// 		element.rank = position++;
+// 	}
+// 	// console.log(arrayData)
+// 	return arrayData;
+// }
 
 export default function BasicTable() {
 	const [selectedDate, setSelectedDate] = useState(today);
@@ -82,7 +82,9 @@ export default function BasicTable() {
 		axios
 			.get('http://localhost:3004/user') //Url api server
 			.then((res) => {
-				setData(sortArray(res.data.data));
+				// setData(sortArray(res.data.data));
+				setData(res.data.data);
+
 			})
 			.catch((error) => {
 				console.log(error);
@@ -142,8 +144,8 @@ export default function BasicTable() {
 			</Toolbar>
 
 			<Divider></Divider>
-			<Table sx={{ minWidth: 650 }} aria-label="simple table">
-				<TableHead>
+			<Table sx={{ minWidth: 650 }} stickyHeader aria-label="table" >
+				<TableHead >
 					<TableRow>
 						<TableCell>Posición</TableCell>
 						<TableCell align="right">Album</TableCell>
