+++ statServer/package.json
@@ -4,7 +4,8 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "start": "tsc && node dist/app.js"
+    "start": "tsc && node dist/app.js",
+    "dev": "tsc && nodemon dist/app.js"
   },
   "keywords": [],
   "author": "",
@@ -21,6 +22,7 @@
     "@types/express": "^4.17.17",
     "@types/mysql": "^2.15.21",
     "@types/node": "^18.15.0",
+    "nodemon": "^2.0.22",
     "typescript": "^4.9.5"
   }
 }
